# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/health/live": {
      "get": {
        "operationId": "HealthController_live",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Health"
        ]
      }
    },
    "/health/ready": {
      "get": {
        "operationId": "HealthController_ready",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Health"
        ]
      }
    },
    "/v1/loan-cases": {
      "post": {
        "operationId": "createCase",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "A caller-generated key, unique per case creation attempt. A retried request with the same key returns the original case rather than creating a second one.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCaseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoanCase"
                }
              }
            }
          },
          "400": {
            "description": "Missing Idempotency-Key header, or a request-body validation failure."
          },
          "401": {
            "description": "Missing or invalid API credentials."
          },
          "404": {
            "description": "jurisdictionCode does not exist."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a loan case (idempotent on Idempotency-Key)",
        "tags": [
          "loan-cases"
        ]
      }
    },
    "/v1/loan-cases/{caseId}": {
      "get": {
        "operationId": "getCase",
        "parameters": [
          {
            "name": "caseId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoanCase"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API credentials."
          },
          "404": {
            "description": "No case with this id owned by the authenticated tenant."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a loan case by id",
        "tags": [
          "loan-cases"
        ]
      }
    },
    "/v1/loan-cases/{caseId}/timeline": {
      "get": {
        "operationId": "getCaseTimeline",
        "parameters": [
          {
            "name": "caseId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TimelineEntry"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API credentials."
          },
          "404": {
            "description": "No case with this id owned by the authenticated tenant."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the chronological event/Agent-run timeline for a case",
        "tags": [
          "loan-cases"
        ]
      }
    },
    "/v1/loan-cases/{caseId}/workflow-runs": {
      "post": {
        "operationId": "startWorkflowRun",
        "parameters": [
          {
            "name": "caseId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartWorkflowRunResult"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API credentials."
          },
          "404": {
            "description": "No case with this id owned by the authenticated tenant."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Start the case-conditions workflow for a case (idempotent per case)",
        "tags": [
          "loan-cases"
        ]
      }
    },
    "/v1/loan-cases/{caseId}/workflow-runs/{runId}": {
      "get": {
        "operationId": "getWorkflowRun",
        "parameters": [
          {
            "name": "caseId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "runId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowRunStatus"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API credentials."
          },
          "404": {
            "description": "No case with this id owned by the authenticated tenant, or no such workflow run."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the status of one workflow run",
        "tags": [
          "loan-cases"
        ]
      }
    },
    "/v1/loan-cases/{caseId}/reviews": {
      "post": {
        "operationId": "submitReview",
        "parameters": [
          {
            "name": "caseId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Signal delivered; the workflow applies it asynchronously."
          },
          "401": {
            "description": "Missing or invalid API credentials."
          },
          "404": {
            "description": "No case with this id owned by the authenticated tenant, or no running workflow for it."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Submit a reviewer decision (condition resolution or evaluation resume)",
        "tags": [
          "loan-cases"
        ]
      }
    },
    "/v1/webhook-endpoints": {
      "post": {
        "operationId": "createWebhookEndpoint",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWebhookEndpointDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookEndpoint"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API credentials."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Register a webhook endpoint. The returned secret is shown only here — no endpoint re-exposes it later.",
        "tags": [
          "webhooks"
        ]
      }
    },
    "/v1/webhook-deliveries/{deliveryId}": {
      "get": {
        "operationId": "getWebhookDelivery",
        "parameters": [
          {
            "name": "deliveryId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookDelivery"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API credentials."
          },
          "404": {
            "description": "No delivery with this id owned by the authenticated tenant."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get one webhook delivery and its full attempt history",
        "tags": [
          "webhooks"
        ]
      }
    }
  },
  "info": {
    "title": "Mortgage Integration Agent — Partner API",
    "description": "Section 15.1 partner REST API. This artifact documents only the endpoints this codebase actually implements today (case creation, lookup, timeline, workflow-run start/status, and reviewer decisions; webhook endpoint registration and delivery lookup) — the full Section 15.1 surface (consents, documents, conditions, evidence, policy-snapshots, provider-operations, audit-export) is not yet built and is not represented here.",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "loan-cases",
      "description": ""
    },
    {
      "name": "webhooks",
      "description": ""
    }
  ],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "A Section 20 M5 scoped API-client credential, \"{clientId}.{secret}\" — minted via `npm run create-api-client` (no self-service REST endpoint exists yet). The tenant is always the one this credential belongs to; no request body or query parameter can specify a different one."
      }
    },
    "schemas": {
      "CreateCaseDto": {
        "type": "object",
        "properties": {
          "borrowerId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "requestedAmount": {
            "type": "number",
            "minimum": 0,
            "exclusiveMinimum": true
          },
          "loanType": {
            "type": "string",
            "enum": [
              "CONVENTIONAL",
              "FHA",
              "VA",
              "JUMBO"
            ]
          },
          "statedMonthlyIncome": {
            "type": "number",
            "minimum": 0,
            "exclusiveMinimum": true
          },
          "jurisdictionCode": {
            "type": "string",
            "minLength": 1,
            "maxLength": 20,
            "example": "US-CA"
          }
        },
        "required": [
          "borrowerId",
          "requestedAmount",
          "loanType",
          "statedMonthlyIncome",
          "jurisdictionCode"
        ]
      },
      "LoanCase": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tenantId": {
            "type": "string",
            "format": "uuid"
          },
          "idempotencyKey": {
            "type": "string"
          },
          "borrowerId": {
            "type": "string"
          },
          "requestedAmount": {
            "type": "number"
          },
          "loanType": {
            "type": "string",
            "enum": [
              "CONVENTIONAL",
              "FHA",
              "VA",
              "JUMBO"
            ]
          },
          "statedMonthlyIncome": {
            "type": "number"
          },
          "jurisdictionCode": {
            "type": "string",
            "example": "US-CA"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "COLLECTING_EVIDENCE",
              "CONDITIONS_OPEN",
              "WAITING_FOR_INFORMATION",
              "WAITING_FOR_REVIEW",
              "READY_FOR_UNDERWRITING",
              "MANUAL_REVIEW",
              "CLOSED"
            ]
          },
          "version": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenantId",
          "idempotencyKey",
          "borrowerId",
          "requestedAmount",
          "loanType",
          "statedMonthlyIncome",
          "jurisdictionCode",
          "status",
          "version",
          "createdAt",
          "updatedAt"
        ]
      },
      "TimelineEntry": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "DOMAIN_EVENT",
              "AGENT_RUN"
            ]
          },
          "summary": {
            "type": "string"
          },
          "detail": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "timestamp",
          "kind",
          "summary",
          "detail"
        ]
      },
      "StartWorkflowRunResult": {
        "type": "object",
        "properties": {
          "workflowId": {
            "type": "string"
          },
          "runId": {
            "type": "string"
          }
        },
        "required": [
          "workflowId",
          "runId"
        ]
      },
      "WorkflowRunStatus": {
        "type": "object",
        "properties": {
          "workflowId": {
            "type": "string"
          },
          "runId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "workflowId",
          "runId",
          "status"
        ]
      },
      "ReviewDto": {
        "type": "object",
        "properties": {
          "reviewType": {
            "type": "string",
            "enum": [
              "CONDITION_RESOLUTION",
              "RESUME_EVALUATION"
            ]
          },
          "actorId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "resolution": {
            "type": "string",
            "enum": [
              "SATISFIED",
              "WAIVED"
            ],
            "description": "Required when reviewType is CONDITION_RESOLUTION."
          },
          "reason": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          }
        },
        "required": [
          "reviewType",
          "actorId"
        ]
      },
      "CreateWebhookEndpointDto": {
        "type": "object",
        "properties": {
          "targetUrl": {
            "type": "string",
            "description": "Where signed delivery attempts are POSTed.",
            "example": "https://partner.example.com/webhooks/mortgage-agent"
          },
          "eventTypes": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "loan_case.created",
                "workflow_run.started",
                "workflow_run.waiting_for_review",
                "workflow_run.completed",
                "workflow_run.failed",
                "evidence.updated",
                "condition.opened",
                "condition.satisfied",
                "condition.waived",
                "evaluation.interrupted",
                "case.escalated",
                "communication.delivered"
              ]
            }
          }
        },
        "required": [
          "targetUrl",
          "eventTypes"
        ]
      },
      "WebhookEndpoint": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tenantId": {
            "type": "string",
            "format": "uuid"
          },
          "targetUrl": {
            "type": "string"
          },
          "secret": {
            "type": "string",
            "description": "The HMAC signing secret for this endpoint. Only ever returned here, at creation — a later read of this row (there is none yet) must not re-expose it."
          },
          "eventTypes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "DISABLED"
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenantId",
          "targetUrl",
          "secret",
          "eventTypes",
          "status",
          "createdAt"
        ]
      },
      "WebhookDeliveryAttempt": {
        "type": "object",
        "properties": {
          "attemptNumber": {
            "type": "number"
          },
          "attemptedAt": {
            "type": "string"
          },
          "httpStatusCode": {
            "type": "object",
            "description": "null on a network-level failure (timeout, connection refused) — no HTTP response was ever received."
          },
          "outcome": {
            "type": "string",
            "enum": [
              "SUCCEEDED",
              "FAILED"
            ]
          },
          "errorMessage": {
            "type": "string"
          }
        },
        "required": [
          "attemptNumber",
          "attemptedAt",
          "httpStatusCode",
          "outcome"
        ]
      },
      "WebhookDelivery": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tenantId": {
            "type": "string",
            "format": "uuid"
          },
          "webhookEndpointId": {
            "type": "string",
            "format": "uuid"
          },
          "outboxEventId": {
            "type": "string",
            "format": "uuid"
          },
          "eventType": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "SUCCEEDED",
              "FAILED_FINAL"
            ]
          },
          "attempts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WebhookDeliveryAttempt"
            }
          },
          "nextAttemptAt": {
            "type": "object"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenantId",
          "webhookEndpointId",
          "outboxEventId",
          "eventType",
          "status",
          "attempts",
          "createdAt"
        ]
      }
    }
  }
}